Fix: multiarch build by moving the binaries build stage to the golang image
Use golang:1.23 as the builder image in the Dockerfile.
With the –platform=$BUILDPLATFORM flag, the image is pulled for the builder host's current architecture.
Cross-compilation occurs in the builder image using TARGETOS and TARGETARCH build arguments to determine the target OS/arch.
These args are set automatically by the multiarch-build process (with docker buildx).
The final container image (registry.access.redhat.com/ubi9/ubi-minimal) is pulled for the correct target architecture, such as amd64 or arm64.

This update fixes support for multiarch builds and speeds up image building, as cross-compilation is faster than compiling on a non-native platform.